TryContextSaveAsync is now returning null and gets catched in Post, Put and Delete

+++ Blob_API/Controllers/ProductController.cs
@@ -117,7 +117,9 @@ namespace Blob_API.Controllers
                     int error = 0;
                     //TODO Update Stock
                     DeleteProductProperty(productToUpdate);
-                    await TryContextSaveAsync();
+                    var result = await TryContextSaveAsync();
+
+                    if (result != null) return result;
                     if (productRessource.Properties != null)
                     {
                        error = await AddPropertyToProduct(productToUpdate, productRessource);
@@ -134,7 +136,9 @@ namespace Blob_API.Controllers
                     }
 
                     DeleteProductCategory(productToUpdate);
-                    await TryContextSaveAsync();
+                    result = await TryContextSaveAsync();
+
+                    if (result != null) return result;
                     if (productRessource.Categories != null)
                     {
                        error = await AddCategoryToProduct(productToUpdate, productRessource);
@@ -146,7 +150,9 @@ namespace Blob_API.Controllers
                     }
 
                     DeleteProductLocation(productToUpdate);
-                    await TryContextSaveAsync();
+                    result = await TryContextSaveAsync();
+
+                    if (result != null) return result;
                     if (productRessource.ProductsAtLocations != null)
                     {
                        error = await AddLocationToProduct(productToUpdate, productRessource);
@@ -164,7 +170,9 @@ namespace Blob_API.Controllers
                     }
                 }
 
-                await TryContextSaveAsync();
+                var result1 = await TryContextSaveAsync();
+
+                if (result1 != null) return result1;
                 await transaction.CommitAsync();
 
                 return NoContent();
@@ -250,7 +258,9 @@ namespace Blob_API.Controllers
                     }
                 }
 
-                await TryContextSaveAsync();
+                var result = await TryContextSaveAsync();
+
+                if (result != null) return result;
                 await transaction.CommitAsync();
 
                 return CreatedAtAction(nameof(GetProductAsync), new {id = newProduct.Id}, newProduct);
@@ -275,11 +285,16 @@ namespace Blob_API.Controllers
                 DeleteProductProperty(productToDelete);
                 DeleteProductCategory(productToDelete);
                 DeleteProductLocation(productToDelete);
-                await TryContextSaveAsync();
+                var result = await TryContextSaveAsync();
+
+                if (result != null) return result;
 
                 _context.Product.Remove(productToDelete);
 
-                await TryContextSaveAsync();
+                result = await TryContextSaveAsync();
+
+                if (result != null) return result;
+
                 await transaction.CommitAsync();
 
                 return NoContent();
@@ -308,7 +323,7 @@ namespace Blob_API.Controllers
                 return Problem("Could not save to Database", statusCode: 500, title: "Error");
             }
 
-            return StatusCode(500);
+            return null;
         }
 
         private async void AddToProductCategoryTable(Product product, Category category)
